Dashboard/backend/requirements.txt
fastapi
uvicorn[standard]
paho-mqtt
sqlalchemy
databases[sqlite]
python-multipart